refactor: `add-dots` > `enhance-dots`
